:construction: (IMUKit): Update interface
