+++ libs/IMUKit/include/IMUKit.hpp
@@ -5,7 +5,7 @@
 #pragma once
 
 #include "interface/LSM6DSOX.hpp"
-#include "interface/libs/IMUKit.h"
+#include "interface/libs/IMUKit.hpp"
 
 namespace leka {
 
@@ -14,21 +14,19 @@ class IMUKit : public interface::IMUKit
   public:
 	explicit IMUKit(interface::LSM6DSOX &lsm6dsox) : _lsm6dsox(lsm6dsox) {}
 
-	void init() final;
+	void init();
 	void start() final;
 	void stop() final;
 
 	void setOrigin() final;
 	void onEulerAnglesReady(angles_ready_callback_t const &callback) final;
-	[[nodiscard]] auto getEulerAngles() const -> interface::EulerAngles final;
+	[[nodiscard]] auto getEulerAngles() const -> EulerAngles final;
 
   private:
 	void drdy_callback(interface::LSM6DSOX::SensorData data);
 
 	interface::LSM6DSOX &_lsm6dsox;
-	interface::EulerAngles _euler_angles
-	{
-	};
+	EulerAngles _euler_angles {};
 	angles_ready_callback_t _on_euler_angles_rdy_callback {};
 };
