+++ libs/IMUKit/source/IMUKit.cpp
@@ -78,7 +78,8 @@ void IMUKit::setOrigin()
 	// ? and start counting from there
 }
 
-auto IMUKit::getEulerAngles() const -> interface::EulerAngles {
+auto IMUKit::getEulerAngles() const -> EulerAngles
+{
 	return _euler_angles;
 }
